Adding delete restriction on images for non owner

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -146,9 +146,17 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes, Model model) {
+        Image image = imageService.getImage(imageId);
+        String error = "Only the owner of the image can delete the image";
+        if (isImageOwner(image, session)) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            redirectAttributes.addAttribute("deleteError", error).addFlashAttribute("deleteError", error);
+            return "redirect:/images/" + image.getId() + "/" + image.getTitle();
+        }
+
     }
